requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
gunicorn==23.0.0
ollama==0.6.1
pydantic==2.13.3
groq==0.28.0
redis==5.0.1
python-dotenv==1.0.1
slowapi==0.1.9
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
